GrammarCells: OptionalCell
You can wrap a cell (or collection) with an "optional".
If it contains a RefNode cell, a rendering condition is generated, so
that the cell is visible, if the child is not set.
A side transformation is generated of of the first cell for creating the
child (works for constants only at the moment).

This is used for the initializer in VariableDeclarations. An initializer
can be added by typing "=".