GrammarCells: Wrapper action for property cells (FloatLiteral)
The wrapper cell can be put around a property cell. If a string is
entered, that is valid for this property, the concept is created with
the entered string as the initial property value.

This is used to create FloatLiterals where an expression is expected.
The string is validated against the regexp of a "constrained string
datatype" and against the "is valid" in the constraints aspect.