GrammarCells: Parentheses can be inserted. Tree is not restructured yet.
There is a new "brackets" cell, that has a "left" and "right" constant
cell and an "inner" RefNode cell for the wrapped node.

You can do a left/right transformation on concepts of the inner cell to
insert the left/right text. It is stored in an annotation.

The next step is to find matching annotations, restructure the tree and
insert the wrapper concept.