GrammarCells: Parser for handling parentheses
There is a new "grammar" cell, that can be used as the root cell for all
concepts, that are allowed to be destructed and reassembled by the
parser.

A "brackets" cell is used to allow the insertion of brackets. The text
in the "left" and "right" cells can be inserted next to subconcepts of
the "inner" cell.

The parser has a timeout to prevent the blocking of the event thread.
For large expressions it can happen, that the parser doesn't find a
valid parse tree in the given time limit.
Some performance improvements have to be done to fix this.