GrammarCells: Support for subconcepts
For example: Actions for PlusExpression are generated from the
BinaryExpression editor.
There is a new constant cell, that allows to implement a query function
for the text based on the subconcept.

The grammar cell allows to specify an additional "projection" cell, that
is used as the concept editor, but not for generating the actions or the
grammar for the parser.