GrammarCells: More parser performance improvement
For each rule the input is split into so many subranges as it contains
symbols. For some rules there are a lot of possible split points.
We can find an invalid split very fast, if we check the constants first.
This is a big performance improvement for long input sequences.