GrammarCells: "+2" side transformation didn't work sometimes
When you want to do a side transformation to a PlusExpression, you type
"+", but because there is also a "+=", the transformation is not
executed. If you then type "2", the side transformation is executed and
the "2" is entered at the new caret position (the right side).
This did not work sometimes, because the caret was placed at the wrong
position.