GrammarCells: Side transformations for multiple OptionCells
In the IfStatement, there are optional "else if"s followed by an
optional "else".
You want to add the "else" after the then block. But if there is an
"else if", it should only be possible to add the else after the else if.

The left and right side transformations are only possible at the
positions where the else or else if block will appear.

It is currently not possible to insert an "else if" between two other
"else if"s. The reason is, that there is a new "side transformation"
cell that allows to specify a side transformation for the exact location
where the cells is added. But this cell cannot be placed between the
elements of a child list. Another cell is required for that.