GrammarCells: Multiple side transformation texts for wrapper cells
The GenericDotExpression allows two side transformations, a dot or an
arrow. Before, there where two wrapper cells in two grammar rules to
make this possible now you can specify the two possible side
transformations in one wrapper cell, which allows it to be moved to the
projection part. The goal is to move all special cells to the projection
part and forbid them in the rule part. This is more consistent.