GrammarCells: Grammar rules are used for showing side transformations of ancestors
The side transformation "return" should be shown on the left side of an
ExpressionStatement. To make this work, you would specify an
"include parent" side transformation on "Expression". In a multi level
expression tree this would result in some duplicate side transformation
(e.g. "+" for all parent expressions).

Now the grammar rules are used to show this side transformation on the
leftmost expression leaf without the requirement for an "include parent".