GrammarCells: Splitting of literals (one string into concatenation of two)
There is a new cell "grammar.splitable" that allows to define a tokenizer
for a property. When you type a string that is not valid for that
property, the parser tries to find a new tree, that matches this string.
Because there is no lexer, you have to manually split the string into
a list of tokens (strings).